docs : 방을 만드는 기능 요구사항 추가
