+++ README.md
@@ -4,6 +4,12 @@
 * 승패 및 점수를 판단할 수 있다.
 * 웹으로 체스 게임이 가능해야 한다.
 * 웹 서버를 재시작하더라도 이전에 하던 체스 게임을 다시 시작할 수 있어야 한다.(DB 적용)
+- 체스방 생성 시 제목과 비밀번호를 입력받는다.
+  이 때 비밀번호는 플레이어1의 비밀번호이다.
+- 체스방 고유식별값 대신 체스방 제목이 표시된다.
+- 참여 시 비밀번호를 입력한다. 
+- 이 때 비밀번호는 플레이어2의 비밀번호이다
+- 화면이 자동으로 갱신될 필요는 없다.
 
 ### 요구사항
 * Spring Framework를 활용하여 애플리케이션을 구동한다.
